Split by label when exporting with filters

+++ neptune-export/src/main/java/com/amazonaws/services/neptune/propertygraph/io/DeletableFile.java
@@ -35,10 +35,12 @@ class DeletableFile implements AutoCloseable {
 
     @Override
     public void close() {
-        boolean deletedOriginalFile = file.delete();
+        if (file.exists()){
+            boolean deletedOriginalFile = file.delete();
 
-        if (!deletedOriginalFile) {
-            throw new IllegalStateException("Unable to delete file: " + file.getAbsolutePath());
+            if (!deletedOriginalFile) {
+                throw new IllegalStateException("Unable to delete file: " + file.getAbsolutePath());
+            }
         }
     }
 }
